Refactor role-based permissions and remove RefreshRolesCommand
Replaced role-based permission checks with a new `@RequiresPermission` annotation for improved clarity and maintainability. Removed `RefreshRolesCommand` and associated manual role-permission handling logic. Updated classes to reflect the new approach and relocated `TheoGuild` to a more appropriate package.
