+++ src/main/java/com/severalcircles/theo/events/MessageEvent.java
@@ -1,12 +1,13 @@
 package com.severalcircles.theo.events;
 
 import com.severalcircles.theo.data.TheoDataManager;
-import com.severalcircles.theo.data.TheoGuild;
+import com.severalcircles.theo.data.guild.TheoGuild;
 import com.severalcircles.theo.data.user.TheoUser;
 import net.dv8tion.jda.api.events.message.MessageReceivedEvent;
 import net.dv8tion.jda.api.hooks.ListenerAdapter;
 
 import java.io.IOException;
+import java.lang.reflect.Member;
 
 public class MessageEvent extends ListenerAdapter {
     @Override
@@ -19,8 +20,11 @@ public class MessageEvent extends ListenerAdapter {
         } catch (IllegalStateException ignored) {}
         TheoGuild g = TheoDataManager.getGuild(event.getGuild().getId());
 
-        TheoDataManager.saveGuild(g);
-        TheoDataManager.saveUser(user);
+
+        if (g != null) TheoDataManager.saveGuild(g);
+        else TheoDataManager.saveGuild(new TheoGuild(user.getGuildId()));
+        if (user != null) TheoDataManager.saveUser(user);
+        else TheoDataManager.saveUser(new TheoUser());
         super.onMessageReceived(event);
     }
 }
