+++ src/main/java/com/severalcircles/theo/events/SlashCommandEvent.java
@@ -2,14 +2,19 @@ package com.severalcircles.theo.events;
 
 import com.severalcircles.theo.Theo;
 import com.severalcircles.theo.data.TheoDataManager;
-import com.severalcircles.theo.data.TheoGuild;
+import com.severalcircles.theo.data.guild.TheoGuild;
 import com.severalcircles.theo.data.user.TheoUser;
+import com.severalcircles.theo.frontend.RequiresPermission;
 import com.severalcircles.theo.frontend.msg.GenericExceptionEmbed;
+import com.severalcircles.theo.frontend.msg.NotAllowedEmbed;
+import net.dv8tion.jda.api.Permission;
 import net.dv8tion.jda.api.events.interaction.command.SlashCommandInteractionEvent;
 import net.dv8tion.jda.api.hooks.ListenerAdapter;
 import org.jetbrains.annotations.NotNull;
 
+import java.util.Arrays;
 import java.util.Locale;
+import java.util.Objects;
 
 public class SlashCommandEvent extends ListenerAdapter {
     @Override
@@ -29,9 +34,18 @@ public class SlashCommandEvent extends ListenerAdapter {
         }
         Theo.commandMap.forEach((k, v) -> {
             if (event.getFullCommandName().equals(k))  {
-                try {v.execute(event, user);}
+                try {
+                    RequiresPermission annotation = v.getClass().getAnnotation(RequiresPermission.class);
+                    if (annotation != null) {
+                        Permission permission = Permission.valueOf(annotation.value());
+                        if (!Objects.requireNonNull(event.getMember()).hasPermission(permission)) {
+                            event.replyEmbeds(new NotAllowedEmbed(Locale.forLanguageTag(user.getLocale()), event.getUser()).get()).queue();
+                            return;
+                        }
+                    }
+                    v.execute(event, user);}
                 catch (Exception e) {
-                    event.replyEmbeds(new GenericExceptionEmbed(Locale.forLanguageTag(user.getLocale()), e).get()).queue();
+                    event.replyEmbeds(new GenericExceptionEmbed(Locale.forLanguageTag("en"), e).get()).queue();
                 }
             }});
         super.onSlashCommandInteraction(event);
