+++ src/main/java/com/severalcircles/theo/frontend/RequiresPermission.java
@@ -1,0 +1,15 @@
+package com.severalcircles.theo.frontend;
+
+import net.dv8tion.jda.api.Permission;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+import java.security.Permissions;
+@Target(ElementType.TYPE)
+@Retention(RetentionPolicy.RUNTIME)
+public @interface RequiresPermission {
+    String value();
+
+}
