+++ src/main/java/com/severalcircles/theo/frontend/basic/BasedCommand.java
@@ -1,14 +1,16 @@
 package com.severalcircles.theo.frontend.basic;
 
 import com.severalcircles.theo.data.user.TheoUser;
+import com.severalcircles.theo.frontend.RequiresPermission;
 import com.severalcircles.theo.frontend.TheoCommand;
 import net.dv8tion.jda.api.events.interaction.command.SlashCommandInteractionEvent;
 
 import java.util.Locale;
-
+@RequiresPermission("MESSAGE_SEND")
 public class BasedCommand implements TheoCommand {
     @Override
     public void execute(SlashCommandInteractionEvent event, TheoUser user) {
         event.replyEmbeds(new BasedEmbed(Locale.getDefault()).get()).queue();
     }
+
 }
