+++ src/main/java/com/severalcircles/theo/frontend/mod/WarnCommand.java
@@ -1,28 +1,26 @@
 package com.severalcircles.theo.frontend.mod;
 
 import com.severalcircles.theo.data.TheoDataManager;
-import com.severalcircles.theo.data.TheoGuild;
+import com.severalcircles.theo.data.guild.TheoGuild;
 import com.severalcircles.theo.data.user.TheoUser;
 import com.severalcircles.theo.data.user.UserWarning;
+import com.severalcircles.theo.frontend.RequiresPermission;
 import com.severalcircles.theo.frontend.TheoCommand;
 import com.severalcircles.theo.frontend.msg.NotAllowedEmbed;
+import net.dv8tion.jda.api.Permission;
 import net.dv8tion.jda.api.entities.User;
 import net.dv8tion.jda.api.events.interaction.command.SlashCommandInteractionEvent;
 
 import java.io.IOException;
-import java.time.OffsetDateTime;
 import java.util.ArrayList;
 import java.util.Date;
 import java.util.Locale;
-
+@RequiresPermission("MESSAGE_MANAGE")
 public class WarnCommand implements TheoCommand {
     @Override
     public void execute(SlashCommandInteractionEvent event, TheoUser user) throws IOException {
         TheoGuild guild = TheoDataManager.getGuild(event.getGuild().getId());
-        if (!user.hasPermission(event.getMember(), new WarnCommand())){
-            event.replyEmbeds(new NotAllowedEmbed(Locale.forLanguageTag(user.getLocale()), event.getUser()).get()).queue();
-            return;
-        }
+
         User target = event.getOption("user").getAsUser();
         TheoUser targetUser = TheoDataManager.getUser(target.getId(), event.getGuild().getId());
         if (targetUser == null) {
